+++ script/main.py
@@ -13,6 +13,8 @@ from run_diagnose import run_diagnose
 from run_fetch_arXiv import run_fetch_arXiv
 from run_fetch_arxiv_from_list import run_fetch_arxiv_from_list
 from run_qiita_api import maybe_update_qiita_from_readme
+from test_command_names import test_command_names
+from test_disabled import test_disabled
 from test_file_content import test_file_content
 from test_is_for_md import test_is_for_md
 from test_mark_down_png import test_mark_down_png
@@ -37,6 +39,8 @@ def main():
     make_web_readme()
     make_web_svg()
     make_enumerate_diagnostics_test()
+    test_command_names()
+    test_disabled()
     test_file_content()
     test_is_for_md()
     test_mark_down_png()